requirements.txt
astropy==5.2.2
astroquery==0.4.6
astrospice==0.2.1
beautifulsoup4==4.12.2
cdflib==0.4.9
celery==5.3.1
dash==2.10.2
dash-ag-grid==2.2.0
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash_daq==0.5.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-mantine-components==0.12.1
dash-table==5.0.0
diskcache==5.6.1
drms
h5netcdf==1.2.0
h5py
HelioPy==0.15.4
ipykernel==6.19.2
ipython==8.12.0
ipython_genutils==0.2.0
ipywidgets==8.0.4
json5==0.9.6
jupyter==1.0.0
jupyter_client==8.1.0
jupyter_console==6.6.3
jupyter_core==5.3.0
jupyter_server==1.23.4
jupyterlab==3.5.3
jupyterlab_pygments==0.1.2
jupyterlab_server==2.22.0
jupyterlab_widgets==3.0.5
lxml
matplotlib==3.7.1
multiprocess
numba==0.56.4
numpy==1.23.5
netCDF4==1.6.3
pandas==1.5.3
Pillow==9.5.0
plotly==5.14.1
plotly_express==0.4.1
seaborn
scipy
spacepy==0.4.1
spiceypy==5.1.2
streamlit==1.23.1
sunpy==4.1.4
tqdm==4.65.0
zeep
